+++ containers/nestjs/src/game/game.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common'
+import { GameGateway } from './game.gateway'
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [GameGateway]
+})
+export class GameModule {}
